Simplify pipeline: remove message_chunks table, chunk in memory
Chunks are transient — no reason to persist them in prod. Instead:
- embed-vectors.js now does everything: fetch → chunk → embed → store
- Track progress via messages.embedded_at column (not a separate table)
- Delete scripts/chunk.js (chunkMessages moved into embed-vectors.js)
- Remove pipeline schema and message_chunks table entirely
- Remove chunk/chunk:prod npm scripts

Messages too short to produce chunks are also marked as embedded so
they don't get re-processed on subsequent runs.
