+++ package.json
@@ -9,8 +9,6 @@
     "download": "node -r dotenv/config scripts/download.js",
     "parse": "node -r dotenv/config scripts/parse.js",
     "parse:prod": "DOTENV_CONFIG_PATH=.env.prod node -r dotenv/config scripts/parse.js",
-    "chunk": "node -r dotenv/config scripts/chunk.js",
-    "chunk:prod": "DOTENV_CONFIG_PATH=.env.prod node -r dotenv/config scripts/chunk.js",
     "embed:vectors": "node -r dotenv/config scripts/embed-vectors.js",
     "embed:vectors:prod": "DOTENV_CONFIG_PATH=.env.prod node -r dotenv/config scripts/embed-vectors.js",
     "fetch-fixture": "node -r dotenv/config scripts/fetch-email-fixture.js",
