+++ scripts/embed-vectors.js
@@ -3,27 +3,27 @@ require('dotenv').config()
 const { parseArgs, DEFAULT_LISTS } = require('./lib/config')
 const { getPool, getSupabase, closePool } = require('./lib/db')
 const { createLogger } = require('./lib/logger')
+const { chunkText } = require('./lib/chunker')
 
 const BATCH_SIZE = 100
 const UPSERT_BATCH_SIZE = 250
 const EMBEDDING_MODEL = 'gte-small'
 
 /**
- * Fetch chunks that haven't been embedded yet
+ * Fetch messages that haven't been embedded yet
  * @param {import('pg').Pool} pool
  * @param {string[]} lists
  * @param {number|null} limit
  * @returns {Promise<Array>}
  */
-async function fetchUnembeddedChunks(pool, lists, limit) {
+async function fetchUnembeddedMessages(pool, lists, limit) {
   const query = `
-    SELECT mc.id, mc.message_id, mc.mailbox_id, mc.chunk_index, mc.chunk_text,
-           m.subject, m.from_email, m.ts
-    FROM pipeline.message_chunks mc
-    JOIN messages m ON m.id = mc.message_id
-    WHERE mc.embedded_at IS NULL
-      AND mc.mailbox_id = ANY($1)
-    ORDER BY m.ts DESC
+    SELECT id, mailbox_id, subject, from_email, ts, body_text
+    FROM messages
+    WHERE body_text IS NOT NULL
+      AND embedded_at IS NULL
+      AND mailbox_id = ANY($1)
+    ORDER BY ts DESC
     ${limit ? `LIMIT ${parseInt(limit, 10)}` : ''}
   `
   const { rows } = await pool.query(query, [lists])
@@ -31,21 +31,50 @@ async function fetchUnembeddedChunks(pool, lists, limit) {
 }
 
 /**
- * Mark chunks as embedded
+ * Mark messages as embedded
  * @param {import('pg').Pool} pool
- * @param {string[]} chunkIds
+ * @param {string[]} messageIds
  */
-async function markChunksEmbedded(pool, chunkIds) {
-  if (chunkIds.length === 0) return
+async function markMessagesEmbedded(pool, messageIds) {
+  if (messageIds.length === 0) return
   await pool.query(
-    `UPDATE pipeline.message_chunks SET embedded_at = now() WHERE id = ANY($1)`,
-    [chunkIds]
+    `UPDATE messages SET embedded_at = now() WHERE id = ANY($1)`,
+    [messageIds]
   )
 }
 
 /**
+ * Process messages into chunks (in memory, no DB)
+ * @param {Array} messages
+ * @returns {Array}
+ */
+function chunkMessages(messages) {
+  const allChunks = []
+
+  for (const msg of messages) {
+    const chunks = chunkText(msg.body_text)
+
+    for (const chunk of chunks) {
+      allChunks.push({
+        id: `${msg.id}#chunk${chunk.index}`,
+        message_id: msg.id,
+        mailbox_id: msg.mailbox_id,
+        chunk_index: chunk.index,
+        chunk_text: chunk.text,
+        token_count: chunk.tokenCount,
+        subject: msg.subject,
+        from_email: msg.from_email,
+        ts: msg.ts,
+      })
+    }
+  }
+
+  return allChunks
+}
+
+/**
  * Build vector objects from chunks and their embeddings
- * @param {Array} chunks - Chunks with metadata from DB
+ * @param {Array} chunks - Chunks with metadata
  * @param {Array<Float32Array|number[]>} embeddings - Embedding arrays
  * @returns {Array} Vector objects for putVectors
  */
@@ -70,11 +99,9 @@ function buildVectors(chunks, embeddings) {
  * @param {Array} chunks
  * @param {Function} embedFn - Function that takes array of texts, returns array of embeddings
  * @param {Object} vectorIndex - Supabase vector bucket index
- * @param {import('pg').Pool} pool
- * @param {Object} logger
  * @returns {Promise<number>} Number of chunks embedded
  */
-async function embedAndStoreBatch(chunks, embedFn, vectorIndex, pool, logger) {
+async function embedAndStoreBatch(chunks, embedFn, vectorIndex) {
   const texts = chunks.map(c => c.chunk_text)
   const embeddings = await embedFn(texts)
 
@@ -86,10 +113,6 @@ async function embedAndStoreBatch(chunks, embedFn, vectorIndex, pool, logger) {
     const { error } = await vectorIndex.putVectors({ vectors: batch })
     if (error) throw new Error(`putVectors failed: ${error.message}`)
   }
-
-  // Mark as embedded in DB
-  const chunkIds = chunks.map(c => c.id)
-  await markChunksEmbedded(pool, chunkIds)
 
   return chunks.length
 }
@@ -120,7 +143,7 @@ async function main() {
   const supabase = getSupabase()
 
   console.log(`\n${'='.repeat(60)}`)
-  console.log(`Embedding chunks for ${lists.length} list(s)`)
+  console.log(`Embedding messages for ${lists.length} list(s)`)
   console.log(`${'='.repeat(60)}`)
 
   try {
@@ -134,34 +157,50 @@ async function main() {
       .from('email-embeddings')
       .index('email-chunks')
 
-    const chunks = await fetchUnembeddedChunks(pool, lists, options.limit)
+    const messages = await fetchUnembeddedMessages(pool, lists, options.limit)
 
-    if (chunks.length === 0) {
-      console.log('\nNo unembedded chunks found.')
+    if (messages.length === 0) {
+      console.log('\nNo unembedded messages found.')
       return
     }
 
-    console.log(`\nFound ${chunks.length} chunks to embed`)
+    console.log(`\nFound ${messages.length} messages to embed`)
 
-    let totalEmbedded = 0
+    let totalChunks = 0
+    let totalSkipped = 0
 
-    for (let i = 0; i < chunks.length; i += BATCH_SIZE) {
-      const batch = chunks.slice(i, i + BATCH_SIZE)
+    for (let i = 0; i < messages.length; i += BATCH_SIZE) {
+      const batch = messages.slice(i, i + BATCH_SIZE)
       const batchNum = Math.floor(i / BATCH_SIZE) + 1
-      const totalBatches = Math.ceil(chunks.length / BATCH_SIZE)
+      const totalBatches = Math.ceil(messages.length / BATCH_SIZE)
 
-      const count = await embedAndStoreBatch(batch, embedFn, vectorIndex, pool, logger)
-      totalEmbedded += count
+      // Chunk in memory
+      const chunks = chunkMessages(batch)
+      const skipped = batch.length - new Set(chunks.map(c => c.message_id)).size
+
+      // Embed and store
+      if (chunks.length > 0) {
+        await embedAndStoreBatch(chunks, embedFn, vectorIndex)
+      }
+
+      // Mark all messages in batch as embedded (including skipped — they don't need re-processing)
+      const batchIds = batch.map(m => m.id)
+      await markMessagesEmbedded(pool, batchIds)
+
+      totalChunks += chunks.length
+      totalSkipped += skipped
 
       process.stdout.write(
-        `\r  Batch ${batchNum}/${totalBatches}: embedded ${count} chunks`
+        `\r  Batch ${batchNum}/${totalBatches}: ${chunks.length} chunks from ${batch.length} messages`
           .padEnd(80)
       )
     }
 
     console.log(`\n\n${'='.repeat(60)}`)
     console.log(`Embedding complete!`)
-    console.log(`  Chunks embedded: ${totalEmbedded}`)
+    console.log(`  Messages processed: ${messages.length}`)
+    console.log(`  Chunks embedded: ${totalChunks}`)
+    console.log(`  Messages skipped (too short): ${totalSkipped}`)
     console.log(`${'='.repeat(60)}`)
   } finally {
     await closePool()
@@ -170,8 +209,9 @@ async function main() {
 
 // Export for testing
 module.exports = {
-  fetchUnembeddedChunks,
-  markChunksEmbedded,
+  fetchUnembeddedMessages,
+  markMessagesEmbedded,
+  chunkMessages,
   buildVectors,
   embedAndStoreBatch,
   EMBEDDING_MODEL,
